Adopt Schema\Implementation and add fromArray for client parsing
